static removed|using unsigned types now

+++ includes/Ship.h
@@ -2,13 +2,13 @@
 
 struct Ship
 {
-	const size_t X;
-	const size_t Y;
-	const short Length;
-	short HP;
-	const short Direction;
+	const uint8_t X;
+	const uint8_t Y;
+	const uint8_t Length;
+	uint8_t HP;
+	const uint8_t Direction;
 
-	Ship(size_t x, size_t y, short len, short dir) : X(x), Y(y), Length(len), Direction(dir), HP(len)
+	Ship(uint8_t x, uint8_t y, uint8_t len, uint8_t dir) : X(x), Y(y), Length(len), Direction(dir), HP(len)
 	{
 
 	}
